+++ R/clean_coordinates.R
@@ -98,7 +98,7 @@
 #' Required to include the 'ranges' test. See \code{\link{cc_iucn}} for details.
 #' @param seas_ref a \code{SpatialPolygonsDataFrame} as alternative reference
 #' for the seas test. If NULL, the 
-#' rnaturalearth::ne_download(scale = 50, type = 'land', category = 'physical') 
+#' rnaturalearth::ne_download(scale = 110, type = 'land', category = 'physical') 
 #' dataset is used.
 #' @param seas_scale The scale of the default landmass reference. Must be one of 10, 50, 110.
 #' Higher numbers equal higher detail. Default = 50.